requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0
